+++ CHANGELOG.md
@@ -17,13 +17,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
   ```rust
   let batch_predictor = batched_fn! { ... };
-  return batch_predictor(input);
+  return batch_predictor(input).await;
   ```
 
   You just need to change the last line to:
 
   ```rust
-  return batch_predictor.evaluate_in_batch(input);
+  return batch_predictor.evaluate_in_batch(input).await;
   ```
 
 ## [v0.2.4](https://github.com/epwalsh/batched-fn/releases/tag/v0.2.4) - 2022-03-14