+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [v0.2.5](https://github.com/epwalsh/batched-fn/releases/tag/v0.2.5) - 2024-03-10
+
 ### Fixed
 
 - Removed unnecessary `Mutex` in `BatchedFn`.